added phone number unique validation
It is done via unconfirmed_phone_number and not directly on phone_number, as the user
never directly interacts with the phone_number property and should never end in a state
where he can not update his profile due to a dangling phone_number that he cant do
anything about.

Also improved feedback in form, showing which phone number is validated, if the user
is trying to update a phone number. Also do not let a user verify an invalid
unconfirmed_phone_number.